Simplified syntax tree definitions
