+++ packages/langium-cli/src/generator/grammar-access-generator.ts
@@ -1,5 +1,5 @@
 import * as langium from 'langium';
-import { AstNode, CompositeGeneratorNode, GeneratorNode, IndentNode, NL, process, findAllFeatures } from 'langium';
+import { CompositeGeneratorNode, GeneratorNode, IndentNode, NL, process, findAllFeatures, getContainerOfType } from 'langium';
 import { LangiumConfig } from '../package';
 
 export function generateGrammarAccess(grammar: langium.Grammar, config: LangiumConfig, bootstrap?: boolean): string {
@@ -66,7 +66,7 @@ function generateFeature(feature: langium.AbstractElement): GeneratorNode {
         indent.children.push("feature: '" + feature.feature + "',", NL);
         indent.children.push("operator: '" + feature.operator + "'", NL);
     } else if (langium.isRuleCall(feature) || langium.isCrossReference(feature) || langium.isKeyword(feature)) {
-        const assignment = <langium.Assignment>AstNode.getContainer(feature, langium.reflection, langium.Assignment);
+        const assignment = getContainerOfType(feature, langium.isAssignment);
         if (assignment) {
             indent.children.push("$type: 'unknown'," , NL, '$container: {', NL, generateAssignment(assignment), '}', NL);
         } else {
