+++ packages/langium/src/default-module.ts
@@ -8,7 +8,7 @@ import { DefaultScopeComputation, DefaultScopeProvider } from './references/scop
 import { DefaultNameProvider } from './references/naming';
 import { ValidationRegistry } from './service/validation/validation-registry';
 import { DefaultDocumentValidator } from './service/validation/document-validator';
-import { AstJsonSerializer } from './service/json-serializer/ast-json-serializer';
+import { DefaultJsonSerializer } from './service/json-serializer/ast-json-serializer';
 
 export type DefaultModuleContext = {
     connection?: Connection
@@ -44,7 +44,7 @@ export function createDefaultModule(context: DefaultModuleContext = {}): Module<
             ValidationRegistry: (injector) => new ValidationRegistry(injector)
         },
         serializer: {
-            JsonSerializer: (injector) => new AstJsonSerializer(injector)
+            JsonSerializer: (injector) => new DefaultJsonSerializer(injector)
         }
     };
 }
