+++ packages/langium/src/index.ts
@@ -21,4 +21,5 @@ export * from './parser/langium-parser';
 export * from './service/completion/content-assist-builder';
 export * from './service/completion/content-assist-service';
 export * from './utils/ast-util';
+export * from './utils/cst-util';
 export * from './utils/stream';
