+++ packages/langium/src/utils/cst-util.ts
@@ -1,0 +1,12 @@
+import { CstNode, LeafCstNode } from '../syntax-tree';
+import { CompositeCstNodeImpl, LeafCstNodeImpl } from '../parser/cst-node-builder';
+
+export function flatten(node: CstNode): LeafCstNode[] {
+    if (node instanceof LeafCstNodeImpl) {
+        return [node];
+    } else if (node instanceof CompositeCstNodeImpl) {
+        return node.children.flatMap(e => flatten(e));
+    } else {
+        return [];
+    }
+}
